Add more links and functions [Delivers #162097398]

+++ static/js/products.js
@@ -16,7 +16,7 @@ function getProducts(){
                     <td>${product.category_type}</td>
                     <td>${product.product_stock}</td>
                     <td>${product.product_specs}</td>
-                    <td>${product.product_price}</td>
+                    <td>${product.product_price}/=</td>
                     <td><input type="button" onclick="getProduct(${product.product_id})" value="View"></td>
                     <td><input type="button" onclick="deleteProduct(${product.product_id})" value="Delete"></td>
                     <td><input type="button" onclick="editPage(${product.product_id})" value="Edit"></td>
@@ -39,11 +39,11 @@ function getProduct(product_id){
         output = "";
             output += `
                 <h2>${data.product_name}</h2>
-                <h3>Product Id: ${data.product_id}</h3>
-                <h3>Category: ${data.category_type}</h3>
-                <h3>Stock: ${data.product_stock}</h3>
-                <h3>Specifications: ${data.product_specs}</h3>
-                <h3>Price: ${data.product_price} /=</h3>
+                <h3>Product Id: <i>${data.product_id}</i></h3>
+                <h3>Category: <i>${data.category_type}</i></h3>
+                <h3>Stock: <i>${data.product_stock}</i></h3>
+                <h3>Specifications: <i>${data.product_specs}</i></h3>
+                <h3>Price: <i>${data.product_price}/=</i></h3>
             `;
             console.log(output);
             sessionStorage.setItem("product", output);
@@ -57,18 +57,25 @@ function deleteProduct(product_id){
     if (user_type == "admin"){
         url = product_url + '/' + product_id;
         console.log(url);
-        fetch(url, {
-            method:'DELETE',
-            headers: access_headers
-        })
-        .then((res) => res.json())
-        .then((data) => {
-            console.log(data);
-            output = `${data.Deleted}`;
-            document.getElementById("message").innerHTML = output;
-        })
-        .then(() =>location.reload())
-        .catch((err) => console.log(err));
+        const confirmation = window.confirm("Are you sure about this action?");
+        if(confirmation){
+            fetch(url, {
+                method:'DELETE',
+                headers: access_headers
+            })
+            .then((res) => res.json())
+            .then((data) => {
+                console.log(data);
+                output = `${data.Deleted}`;
+                document.getElementById("message").innerHTML = output;
+            })
+            .then(() =>location.reload())
+            .catch((err) => console.log(err));
+        }
+        else{
+            alert("Delete has been cancled!");
+        }
+        
     }
     else{
         alert("Only Admin can Perform Delete products");
